+++ pybamm/solvers/dae_solver.py
@@ -227,10 +227,10 @@ class DaeSolver(pybamm.BaseSolver):
             initial_conditions
         """
         # Convert model attributes to casadi
-        t_casadi = casadi.SX.sym("t")
+        t_casadi = casadi.MX.sym("t")
         y0 = model.concatenated_initial_conditions
-        y_diff = casadi.SX.sym("y_diff", len(model.concatenated_rhs.evaluate(0, y0)))
-        y_alg = casadi.SX.sym(
+        y_diff = casadi.MX.sym("y_diff", len(model.concatenated_rhs.evaluate(0, y0)))
+        y_alg = casadi.MX.sym(
             "y_alg", len(model.concatenated_algebraic.evaluate(0, y0))
         )
         y_casadi = casadi.vertcat(y_diff, y_alg)