+++ tests/integration/test_models/test_full_battery_models/test_lead_acid/test_side_reactions/test_full_side_reactions.py
@@ -44,7 +44,7 @@ class TestLeadAcidFullSideReactions(unittest.TestCase):
         model = pybamm.lead_acid.Full(options)
         parameter_values = model.default_parameter_values
         parameter_values.update(
-            {"Current function": pybamm.GetConstantCurrent(current=0)}
+            {"Current function": pybamm.ConstantCurrent(current=0)}
         )
         modeltest = tests.StandardModelTest(model, parameter_values=parameter_values)
         modeltest.test_all(skip_output_tests=True)